covid_data

A simulated dataset containing disease transmission parameters and outbreak risk measures for different COVID-19 variants. This dataset supports the COVID-19 Outbreak Risk Explorer Shiny app included in the package.

covid_data

Format

A data frame with 60 rows and 7 variables:

variant

COVID-19 variant (e.g., Wildtype, Delta, Omicron).

vaccination_rate

Proportion of the population vaccinated.

vaccine_efficacy

Effectiveness of the vaccine in preventing infection.

testing_freq_days

Average number of days between population testing.

R0

Basic reproduction number — the average number of secondary infections caused by one infected person.

outbreak_risk

Simulated probability of a COVID-19 outbreak under given conditions.

X

Row identifier used internally (not relevant for analysis).

Source

Simulated dataset based on model parameters described in Zachreson et al. (2021), "A model of COVID-19 quarantine and outbreak risk in Australia".

Details

The dataset was created to demonstrate how disease transmission parameters interact to affect outbreak potential across variants. It is used within the Shiny app CovidDataPackage1::run_app() to generate interactive visualizations.

Examples

data("covid_data", package = "CovidDataPackage1")
head(covid_data)
#>    variant vaccination_rate vaccine_efficacy testing_freq_days       R0
#> 1 Wildtype             0.55             0.64                 7 2.611244
#> 2    Delta             0.32             0.67                 3 5.124853
#> 3 Wildtype             0.68             0.87                 7 2.421885
#> 4 Wildtype             0.98             0.85                 3 2.407413
#> 5    Delta             0.62             0.60                 3 5.334972
#> 6    Delta             0.21             0.84                 5 5.955254
#>   outbreak_risk
#> 1         0.142
#> 2         0.145
#> 3         0.083
#> 4         0.014
#> 5         0.121
#> 6         0.294